feat: Add goBack functionality to SidebarProvider and Navbar
This commit adds the goBack functionality to the SidebarProvider and Navbar components. When the user selects the "goBack" option, the command "workbench.action.navigateBack" is executed, allowing the user to navigate back to the previous page. This feature improves the user experience by providing an easy way to navigate within the extension.